using memoized user now
